refactor: 💡 simplify transformer calls
