+++ src/processors/globalStyle.ts
@@ -3,9 +3,7 @@ import { PreprocessorGroup } from '../types';
 export default (): PreprocessorGroup => {
   return {
     async style({ content, attributes, filename }) {
-      const { default: transformer } = await import(
-        '../transformers/globalStyle'
-      );
+      const { transformer } = await import('../transformers/globalStyle');
 
       if (!attributes.global) {
         return { code: content };
