+++ src/processors/replace.ts
@@ -2,7 +2,7 @@ import { PreprocessorGroup, Options } from '../types';
 
 export default (options: Options.Replace): PreprocessorGroup => ({
   async markup({ content, filename }) {
-    const { default: transformer } = await import('../transformers/replace');
+    const { transformer } = await import('../transformers/replace');
 
     return transformer({ content, filename, options });
   },
